修复修改流水记录后崩溃的BUG；修复删除标签后崩溃的BUG；修正与编号相关的数据类型以符合实际用法

+++ app/src/main/java/com/project/manager/ui/bookkeeping/flow_edit/new_flow/NewFlowActivity.java
@@ -98,8 +98,8 @@ public class NewFlowActivity extends AppCompatActivity implements View.OnClickLi
         String flowRemark = currentFragment.getRemark();    //备注
         dataBundle.putString(KeyValueStrings.FLOW_REMARK.getValue(), flowRemark);
         String tag_name = currentFragment.getFlowTag();     //标签
-        int tag_no = Tag.nameTransToTno(tag_name, this);
-        dataBundle.putInt(KeyValueStrings.TAG_NO.getValue(), tag_no);
+        long tag_no = Tag.nameTransToTno(tag_name, this);
+        dataBundle.putLong(KeyValueStrings.TAG_NO.getValue(), tag_no);
 
         //获取碎片特殊信息并打包
         if (flowType == FlowTypeEnum.TRANSFER) {
